[fr] fix 1100, add new prefix "informal" for 1100–1900
and move "onze-cents" to there, so 1100 is "mille-cent"
now, as preferred in the written language, and in particular
in a legal, administrative or scientific text, according
to Dictionnaire de l’Académie française, 9ᵉ édition:

https://www.dictionnaire-academie.fr/article/A9M2155#:~:text=Dans%20l'usage%20courant%2C%20au,cents%20francs%2C%20seize%20cents%20euros.

Note: 1200 was already mille-deux-cents, but now
with "informal", "douze-cents", etc. up to "dix-neuf-cents".

Reported by "arena94" in GitHub issue #90.